% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_fun.R
\name{CI_fun}
\alias{CI_fun}
\title{Confidence Interval for the Spatial Dependence Parameter}
\usage{
CI_fun(p, rhoHat, adj_mat)
}
\arguments{
\item{p}{Selected truncation level.}

\item{rhoHat}{Estimate of the spatial dependence parameter \eqn{\rho}.}

\item{adj_mat}{An binary adjagency matrix defining the spatial neighborhood structure.}
}
\value{
A numeric vector of length 2 giving the lower and upper bounds of the 95\% confidence interval for \eqn{\rho}.
}
\description{
Computes the 95\% confidence interval for the spatial dependence parameter \eqn{\rho}.
}
\examples{

library(FCARfreq)

# Load example data included in the package
data(data_example)
data <- data_example$data
nbd_index <- data_example$nbd_index

# Create an adjacency matrix from nbd_index
n <- nrow(data)
adj_mat <- matrix(0, nrow = n, ncol = n)
for(i in 1:n){
 adj_mat[i, nbd_index[[i]]] <- 1
}

# Estimation of alpha
alphaHat <- colMeans(data)

 est_FVE95 <- FCAR_est_fun(alphaHat, rhoHat=0.3, adj_mat, nbd_index, data, sel_p="FVE95")
 
CI_fun(est_FVE95$p, est_FVE95$rhoHat, adj_mat)

}
\seealso{
\code{\link{FCAR_est_fun}}
}
