% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_example.R
\docType{data}
\name{data_example}
\alias{data_example}
\title{Example Dataset: data_example}
\format{
A list with 2 elements:
\describe{
\item{\code{data}:}{A simulated spatial functional dataset consisting of 900 curves observed on 30*30 lattice. 
Each curve is evaluated at 50 equally spaced points over the interval [0, 1].}
\item{\code{nbd_index}:}{An \code{nb} class object specifying the neighborhood structures, 
in which each element lists the adjacent spatial locations.}
}
}
\usage{
data(data_example)
}
\description{
This is a simulated dataset on 30*30 lattice.
}
\keyword{datasets}
