% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCAR_est_fun.R
\name{FCAR_est_fun}
\alias{FCAR_est_fun}
\title{Estimation of Functional Conditional Autoregressive (FCAR) Model}
\usage{
FCAR_est_fun(
  alphaHat,
  rhoHat,
  adj_mat,
  nbd_index,
  data,
  sel_p = c("FVE75", "FVE80", "FVE85", "FVE90", "FVE95", "FVE99"),
  rho_tol = 1e-08,
  G_tol = 1e-06
)
}
\arguments{
\item{alphaHat}{Estimated value of \eqn{\alpha}, the mean function of the joint distribution.}

\item{rhoHat}{An initial value of \eqn{\rho}, the spatial dependence parameter.}

\item{adj_mat}{An binary adjagency matrix defining the spatial neighborhood structure.}

\item{nbd_index}{An nb class object specifying the neighborhood structures, in which each element lists the adjacent spatial locations.}

\item{data}{An n by T matrix of functional data. Each row represents one functional observation evaluated on a common grid.}

\item{sel_p}{A character string specifying the threshold level for the fraction of variance explained (FVE) used to select the truncation level. 
Must be one of \code{"FVE75"}, \code{"FVE80"}, \code{"FVE85"}, \code{"FVE90"}, \code{FVE95}, or \code{"FVE99"}.}

\item{rho_tol}{Convergence tolerance for the spatial dependence parameter \eqn{\rho}.}

\item{G_tol}{Convergence tolerance for the covariance operator.}
}
\value{
A list with the following elements:
\describe{
 \item{\code{rhoHat}:}{Estimate of the spatial dependence parameter \eqn{\rho}.}
 \item{\code{GHat}:}{Estimate of covariance operator.}
 \item{\code{iter}:}{Number of iterations used in the profile-based estimation procedure.}
 \item{\code{p}:}{Selected truncation level.}
}
}
\description{
Implements the profile-based estimation procedure described in Algorithm 1 of the main paper.
}
\examples{

library(FCARfreq)

# Load example data included in the package
data(data_example)
data <- data_example$data
nbd_index <- data_example$nbd_index

# Create an adjacency matrix from nbd_index
n <- nrow(data)
adj_mat <- matrix(0, nrow = n, ncol = n)
for(i in 1:n){
 adj_mat[i, nbd_index[[i]]] <- 1
}

# Estimation of alpha
alphaHat <- colMeans(data)

 est_FVE95 <- FCAR_est_fun(alphaHat, rhoHat=0.3, adj_mat, nbd_index, data, sel_p="FVE95")
 
# Example: estimated spatial dependence parameter
est_FVE95$rhoHat

}
\seealso{
\code{\link{CI_fun}}
}
