\name{FCARfreq-package}
\alias{FCARfreq-package}
\alias{FCARfreq}
\docType{package}
\title{
\packageTitle{FCARfreq}
}
\description{
\packageDescription{FCARfreq}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{FCARfreq}

Maintainer: \packageMaintainer{FCARfreq}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = FCARfreq)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{FCARfreq}}
%\section{Documentation Index}{\packageIndices{FCARfreq}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
